deprecated/algorithms-deprec/requirements.txt
# Core ML and Data Processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
torch>=1.9.0

# NLP and Text Processing
nltk>=3.6.3
transformers>=4.11.0
spacy>=3.1.0

# Database and Caching
redis>=3.5.3
psycopg2-binary>=2.9.1

# Data Models and Validation
pydantic>=1.8.2

# Utilities
python-dotenv>=0.19.0
loguru>=0.5.3
tqdm>=4.62.2 